+++ internal/cchook/skiplist.go
@@ -1,5 +1,0 @@
-package cchook
-
-func SkipList() []string {
-	return []string{"cd", "false", "git status", "htop", "less", "ls", "man", "more", "nano", "nvim", "opsmask", "pwd", "top", "true", "vi", "vim"}
-}